Adopt commit 4267e30fa from master: Merge pull request #1178 from jmendelsohn4/drqs-101945678/balxml_nb_20170609
commit 4267e30fafdfff5232ebc9a253141f6f819cf514
Merge: d8c244e41 39fc6409e
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Fri Jun 9 13:16:28 2017 -0400

    Merge pull request #1178 from jmendelsohn4/drqs-101945678/balxml_nb_20170609

    clean test driver warnings

commit 39fc6409e1d761797478e33741a7ee98492fbe14
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Fri Jun 9 10:39:04 2017 -0400

    clean test driver warnings

+++ groups/bal/balxml/balxml_encodingstyle.t.cpp
@@ -186,7 +186,9 @@ int main(int argc, char *argv[])
                 const char *STR       = DATA[ti].d_inputStr;
 
                 Enum value;
-                int rc = Obj::fromString(&value, STR, strlen(STR));
+                int rc = Obj::fromString(&value,
+                                         STR,
+                                         static_cast<int>(strlen(STR)));
 
                 if (veryVerbose) { T_; P_(ti); P_(value); P_(STR); P(rc); }
